updated git modules, packages, README and lint (#25)
* updated forge-std and ds-test .gitmodule.
* added chainlink node modules.
* updated README.md for deployment scripts.
* created deployment scripts to deploy ChainlinkInvertingOracle.sol.
* updated Helper.sol to replace some constant values.
* ran lint and prettier.

+++ src/ChainlinkOracle.sol
@@ -10,14 +10,13 @@ contract ChainlinkOracle is IOracle {
 
     constructor(address _dataFeedAddress, uint256 _decimals) {
         dataFeed = AggregatorV3Interface(_dataFeedAddress);
-        scalingFactor = 10 ** (uint256(dataFeed.decimals()) - _decimals);
+        scalingFactor = 10**(uint256(dataFeed.decimals()) - _decimals);
     }
 
     function acceptTermsOfService() external {}
 
     function readData() external view returns (uint256) {
-
-        (, int answer,,,) = dataFeed.latestRoundData();
+        (, int256 answer, , , ) = dataFeed.latestRoundData();
         return (uint256(int256(answer)) / scalingFactor);
     }
 }
